Добавлен отдельный класс шагов ApiSteps для апи реквеста, этот шаг удален из AuthorizationSteps
